Refactor Parser file
