+++ lib/fine-types/src/Language/FineTypes/Parser/Types.hs
@@ -1,0 +1,7 @@
+module Language.FineTypes.Parser.Types where
+
+import Data.String (String)
+import Data.Void (Void)
+import Text.Megaparsec (Parsec)
+
+type Parser = Parsec Void String
